requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
nltk==3.8.1
pdfplumber==0.11.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
trafilatura==1.6.0
pymupdf
pytesseract
pdf2image